Add files via upload
The main idea behind this project is to display the Traffic Signal with some  cars, sky, plane, cloud, pond, house, tree and other scenario with Computer Graphics.
In this  project , when we keyboard buttons  press ‘g’or ‘G' to change the signal light to Green then the cars are moving one side to another side. In this  project , when we keyboard buttons  press ‘y' or ‘Y' to change the signal light to Yellow then the cars are slowly moving one side to another side. And when we press 'r' or 'R' to change the signal light to Red then the cars are stopped. In this  project ,we want to see in the day scenario when we press 'd' or ‘D’ or we want to see in the night scenario so we press 'n' or 'N’. Press RIGHT MOUSE BUTTON To Display Menu : Select  'Aeroplane' to add moving Aeroplane.It’s moving in the day and night are both time. Select  'Comet' to add moving Comet.In the night just it’s moving one side to another side. Select  'Quite' to Exit the application.